Backend/requirements.txt
fastapi==0.129.0
uvicorn==0.41.0
sqlalchemy==2.0.46
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic-settings==2.13.1
python-dotenv==1.2.1
alembic==1.18.4
pytest==8.4.2
httpx==0.28.1
passlib[bcrypt]==1.7.4
bcrypt==4.1.3
python-jose[cryptography]==3.5.0
Authlib==1.6.5
email-validator==2.3.0
